L5pc debug

+++ examples/l5pc/opt_l5pc.py
@@ -24,17 +24,15 @@ This optimisation is based on L5PC optimisations developed by Etay Hay in the
 context of the BlueBrain project
 """
 
+# pylint: disable=R0914, W0403
+import os
 
-# pylint: disable=R0914
-
-import bluepyopt
-
-# TODO store definition dicts in json
-# TODO rename 'score' into 'objective'
-# TODO add functionality to read settings of every object from config format
+import argparse
 
 import l5pc_evaluator
 evaluator = l5pc_evaluator.create()
+
+import bluepyopt
 
 
 def evaluate(parameter_array):
@@ -51,16 +49,41 @@ opt = bluepyopt.optimisations.DEAPOptimisation(
 def main():
     """Main"""
 
-    import argparse
     parser = argparse.ArgumentParser(description='L5PC example')
     parser.add_argument('--start', action="store_true")
     parser.add_argument('--continue_cp', action="store_true")
     parser.add_argument('--analyse', action="store_true")
+    parser.add_argument('--compile', action="store_true")
+    parser.add_argument('--hocanalyse', action="store_true")
+
+    args = parser.parse_args()
+
+    if args.compile:
+        import commands
+        commands.getstatusoutput('cd mechanisms/; nrnivmodl; cd ..')
+
+    # TODO store definition dicts in json
+    # TODO rename 'score' into 'objective'
+    # TODO add functionality to read settings of every object from config format
+
+    if args.hocanalyse:
+        try:
+            import bglibpy  # NOQA
+        except ImportError:
+            raise ImportError(
+                'bglibpy not installed, '
+                '--hocanalyse for internal testing only!')
+    else:
+        mechfile = "./mechanisms/x86_64/.libs/libnrnmech.so"
+        if os.path.isfile(mechfile):
+            from bluepyopt.ephys import neuron
+            neuron.h.nrn_load_dll(mechfile)
+
+        else:
+            raise ImportError('nrnmech not compiled, run --compile first!')
 
     # TODO read checkpoint filename from arguments
     cp_filename = 'checkpoints/checkpoint.pkl'
-
-    args = parser.parse_args()
 
     if args.start or args.continue_cp:
         opt.run(
@@ -89,17 +112,47 @@ def main():
 
         fig_release.savefig('figures/release_l5pc.eps')
 
-        bpop_model_fig = plt.figure(figsize=(10, 10), facecolor='white')
-        bpop_evol_fig = plt.figure(figsize=(10, 10), facecolor='white')
+        if os.path.isfile(cp_filename):
 
-        l5pc_analysis.analyse_cp(
+            bpop_model_fig = plt.figure(figsize=(10, 10), facecolor='white')
+            bpop_evol_fig = plt.figure(figsize=(10, 10), facecolor='white')
+
+            l5pc_analysis.analyse_cp(
+                opt=opt,
+                cp_filename=cp_filename,
+                figs=[bpop_model_fig, bpop_evol_fig],
+                boxes=[box, box])
+
+            bpop_model_fig.savefig('figures/bpop_l5pc_model.eps')
+            bpop_evol_fig.savefig('figures/bpop_l5pc_evolution.eps')
+
+        else:
+            print('No checkpoint file available run optimization '
+                  'first with --start')
+
+        plt.show()
+
+    elif args.hocanalyse:
+
+        import l5pc_analysis
+
+        # _, axes_obj = plt.subplots(n_of_rows, n_of_cols, facecolor='white')
+        # axes = numpy.ravel(axes_obj)
+        import matplotlib.pyplot as plt
+        fig_release = plt.figure(figsize=(10, 10), facecolor='white')
+
+        box = {
+            'left': 0.0,
+            'bottom': 0.0,
+            'width': 1.0,
+            'height': 1.0}
+
+        l5pc_analysis.analyse_releasecircuit_hocmodel(
             opt=opt,
-            cp_filename=cp_filename,
-            figs=[bpop_model_fig, bpop_evol_fig],
-            boxes=[box, box])
+            fig=fig_release,
+            box=box)
 
-        bpop_model_fig.savefig('figures/bpop_l5pc_model.eps')
-        bpop_evol_fig.savefig('figures/bpop_l5pc_evolution.eps')
+        fig_release.savefig('figures/release_l5pc_hoc.eps')
 
         plt.show()
